Поменял сущности для лучшей совместимости

+++ Domain/Entities/PersonInContent.cs
@@ -4,7 +4,7 @@ public class PersonInContent
 {
     public int Id { get; set; }
 
-	public int ContentId { get; set; }
+	public long ContentId { get; set; }
 	public ContentBase Content { get; set; }
 
     public string Name { get; set; } = null!;
